[Step2] 테바(김성겸) JDBC 구현하기 미션 제출합니다. (#858)
* refactor: 불필요 필드 제거

* refactor: 로깅 추가

* refactor: 불필요 메서드 제거

* test: RowMapperTest 버그 픽스

* refactor: 쿼리 실행 시 반복되는 try catch 문 추출

+++ jdbc/src/test/java/com/interface21/jdbc/core/RowMapperTest.java
@@ -16,13 +16,13 @@ class RowMapperTest {
     void mapRow() throws SQLException {
         ResultSet resultSet = mock(ResultSet.class);
 
-        when(resultSet.getInt("id")).thenReturn(1);
+        when(resultSet.getLong("id")).thenReturn(1L);
         when(resultSet.getString("name")).thenReturn("John Doe");
 
         RowMapper<TestUser> rowMapper = new TestUserRowMapper();
         TestUser testUser = rowMapper.mapRow(resultSet, 1);
 
-        assertThat(testUser.getId()).isEqualTo(1);
+        assertThat(testUser.getId()).isEqualTo(1L);
         assertThat(testUser.getName()).isEqualTo("John Doe");
     }
 }
